--- gpt3chatbot/_utils.py
@@ -1,15 +1,0 @@
-import functools
-
-
-def memoize(obj):
-    """Decorator for memoizing functions from https://wiki.python.org/moin/PythonDecoratorLibrary#Memoize"""
-    cache = obj.cache = {}
-
-    @functools.wraps(obj)
-    def memoizer(*args, **kwargs):
-        key = str(args) + str(kwargs)
-        if key not in cache:
-            cache[key] = obj(*args, **kwargs)
-        return cache[key]
-
-    return memoizer
